+++ api/v0042/api.go
@@ -1,0 +1,8 @@
+// SPDX-FileCopyrightText: Copyright (C) SchedMD LLC.
+// SPDX-License-Identifier: Apache-2.0
+
+package v0042
+
+// Ref: https://slurm.schedmd.com/rest_api.html
+
+//go:generate go run github.com/oapi-codegen/oapi-codegen/v2/cmd/oapi-codegen@v2.4.1 -config oapi-codegen-config.yaml slurm-openapi.gen.json