+++ CHANGELOG.md
@@ -6,6 +6,8 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 
+- Added v42 OpenAPI spec and generated API.
+
 ### Fixed
 
 ### Changed